+++ R/gadget.R
@@ -131,7 +131,7 @@ gitgadget <- function(port = get_port()) {
           ),
           conditionalPanel("input.create_remote == 'GitLab'",
             fillRow(height = "70px", width = "300px",
-              textInput("create_group","Group name:", value = Sys.getenv("git.group")),
+              textInput("create_group", "Group name:", value = Sys.getenv("git.group")),
               uiOutput("ui_create_pre")
             )
           ),
@@ -154,6 +154,8 @@ gitgadget <- function(port = get_port()) {
           actionButton("remove_git_show", "Remove .git", title = "Remove local .git directory if present", class = "btn-danger"),
           HTML("<h4>Create local .git and remote repo</h4>"),
           actionButton("create", "Create", title = "Create a new repo using the gitlab API"),
+          actionButton("create_hide_repo", "Hide", title = "Hide class repo from students", class = "btn-warning"),
+          actionButton("create_show_repo", "Show", title = "Show class repo to students", class = "btn-success"),
           hr(),
           verbatimTextOutput("create_output")
         )
@@ -237,8 +239,10 @@ gitgadget <- function(port = get_port()) {
               ),
               textInput("collect_server","API server:", value = Sys.getenv("git.server", "https://gitlab.com/api/v4/")),
               radioButtons("collect_type", "Assignment type:", c("individual","team"), "individual", inline = TRUE),
+              actionButton("collect_show_repo", "Show", title = "Show class repo to students", class = "btn-success"),
               actionButton("collect", "Collect", title = "Create merge requests from all student forks using the gitlab API. Used for assignment management by instructors"),
-              actionButton("collect_fetch", "Fetch", title = "Create local branches from all merge requests and link them to (new) remote branches. Used for assignment management by instructors")
+              actionButton("collect_fetch", "Fetch", title = "Create local branches from all merge requests and link them to (new) remote branches. Used for assignment management by instructors"),
+              actionButton("collect_hide_repo", "Hide", title = "Hide class repo from students", class = "btn-warning")
             ),
             hr(),
             verbatimTextOutput("collect_output")
@@ -1381,6 +1385,60 @@ gitgadget <- function(port = get_port()) {
       })
 
       message("\nUse the Git tab in R-studio (click refresh first) to switch between different student assignment submissions\n")
+    })
+
+    create_repo_name <- reactive({
+
+      repo <- basename(input$create_directory)
+
+      create_pre_lc <- tolower(input$create_pre)
+      if (!is_empty(create_pre_lc)) {
+        repo <- paste0(create_pre_lc, repo)
+      }
+
+      create_group_lc <- tolower(input$create_group)
+      if (!is_empty(create_group_lc) && create_group_lc != Sys.getenv("git.user")) {
+        repo <- paste0(create_group_lc, "/", repo)
+      }
+
+      repo
+
+    })
+
+    observeEvent(input$create_hide_repo, {
+      req(input$create_token, input$create_server, input$create_user_file)
+
+      withProgress(message = "Hiding class repo", value = 0, style = "old", {
+        repo <- create_repo_name()
+        remove_users_repo(input$create_token, repo, input$create_user_file, server = input$create_server)
+        cat("\nUser permissions removed ...\n\n")
+      })
+    })
+
+
+    observeEvent(input$create_show_repo, {
+      req(input$create_token, input$create_server, input$create_user_file)
+      withProgress(message = "Showing class repo", value = 0, style = "old", {
+        repo <- create_repo_name()
+        add_users_repo(input$create_token, repo, input$create_user_file, permission = 20, server = input$create_server)
+        cat("User permissions added ...\n")
+      })
+    })
+
+    observeEvent(input$collect_hide_repo, {
+      req(input$collect_token, input$collect_server, input$collect_user_file, input$collect_assignment)
+      withProgress(message = "Hiding class repo", value = 0, style = "old", {
+        remove_users_repo(input$collect_token, input$collect_assignment, input$collect_user_file, server = input$collect_server)
+        cat("\nUser permissions removed ...\n")
+      })
+    })
+
+    observeEvent(input$collect_show_repo, {
+      req(input$collect_token, input$collect_server, input$collect_user_file, input$collect_assignment)
+      withProgress(message = "Showing class repo", value = 0, style = "old", {
+        add_users_repo(input$collect_token, input$collect_assignment, input$collect_user_file, permission = 20, server = input$create_server)
+        cat("User permissions added ...\n\n")
+      })
     })
 
     output$collect_output <- renderPrint({